Refactor to file scoped namespace, adjust namespace and rename AgentSettings to IntegrationSettings
