+++ Integration.Common/Classes/BaseIntegration.cs
@@ -2,21 +2,20 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 
-namespace Integration.Common.Classes
+namespace Integration.Common.Classes;
+
+public abstract class BaseIntegration : IIntegration
 {
-    public abstract class BaseIntegration : IIntegration
+    public IOptions<IntegrationSettings> Options { get; }
+
+    public BaseIntegration(IOptions<IntegrationSettings> options)
     {
-        public IOptions<AgentSettings> Options { get; }
-
-        public BaseIntegration(IOptions<AgentSettings> options)
-        {
-            Options = options;
-        }
-
-        public abstract Guid Id { get; set; }
-        public abstract string Name { get; set; }
-        public abstract void RegisterServices(IServiceCollection serviceCollection);
-        public abstract Task StartAsync(CancellationToken cancellationToken);
-        public abstract Task StopAsync(CancellationToken cancellationToken);
+        Options = options;
     }
+
+    public abstract Guid Id { get; set; }
+    public abstract string Name { get; set; }
+    public abstract void RegisterServices(IServiceCollection serviceCollection);
+    public abstract Task StartAsync(CancellationToken cancellationToken);
+    public abstract Task StopAsync(CancellationToken cancellationToken);
 }
