+++ Integration.Common/Notifications/IntegrationOutputFileCreated.cs
@@ -1,14 +1,13 @@
 ﻿using MediatR;
 
-namespace Integration.Common.Notifications
-{
-    public class IntegrationOutputFileCreated : INotification
-    {
-        public string FilePath { get; }
+namespace Integration.Common.Notifications;
 
-        public IntegrationOutputFileCreated(string filePath)
-        {
-            FilePath = filePath;
-        }
+public class IntegrationOutputFileCreated : INotification
+{
+    public string FilePath { get; }
+
+    public IntegrationOutputFileCreated(string filePath)
+    {
+        FilePath = filePath;
     }
 }
