+++ Integration.Common/Requests/SendMessageWithAgent.cs
@@ -5,58 +5,57 @@ using MetalHeaven.Agent.Shared.External.Helpers;
 using MetalHeaven.Agent.Shared.External.Interfaces;
 using Microsoft.Extensions.Options;
 
-namespace Integration.Common.Requests
+namespace Integration.Common.Requests;
+
+public class SendMessageWithAgent : IRequest
 {
-    public class SendMessageWithAgent : IRequest
+    public class Request : IRequest<Response>
     {
-        public class Request : IRequest<Response>
+        public Request(IAgentMessage message, string integrationName)
         {
-            public Request(IAgentMessage message, string integrationName)
-            {
-                Message = message;
-                IntegrationName = integrationName;
-            }
-
-            public IAgentMessage Message { get; }
-            public string IntegrationName { get; }
+            Message = message;
+            IntegrationName = integrationName;
         }
 
-        public class Response
+        public IAgentMessage Message { get; }
+        public string IntegrationName { get; }
+    }
+
+    public class Response
+    {
+    }
+
+    public class Handler : IRequestHandler<Request, Response>
+    {
+        private readonly IntegrationSettings _settings;
+
+        public Handler(IOptions<IntegrationSettings> options)
         {
+            _settings = options.Value;
         }
 
-        public class Handler : IRequestHandler<Request, Response>
+        public async Task<Response> Handle(Request request, CancellationToken cancellationToken)
         {
-            private readonly AgentSettings _settings;
+            // define temp file path
+            var tempFilePath = Path.GetTempFileName();
 
-            public Handler(IOptions<AgentSettings> options)
-            {
-                _settings = options.Value;
-            }
+            // convert message to json
+            var json = request.Message.ToJson();
 
-            public async Task<Response> Handle(Request request, CancellationToken cancellationToken)
-            {
-                // define temp file path
-                var tempFilePath = Path.GetTempFileName();
+            // write json to temp file
+            await File.WriteAllTextAsync(tempFilePath, json, cancellationToken).ConfigureAwait(false);
 
-                // convert message to json
-                var json = request.Message.ToJson();
+            // define agent input directory
+            var agentInputDirectory = _settings.GetOrCreateAgentInputDirectory(request.IntegrationName, true);
 
-                // write json to temp file
-                await File.WriteAllTextAsync(tempFilePath, json, cancellationToken).ConfigureAwait(false);
+            // define new path to move temp file to
+            var destinationFilePath = Path.Combine(agentInputDirectory, $"{Guid.NewGuid()}.json");
 
-                // define agent input directory
-                var agentInputDirectory = _settings.GetOrCreateAgentInputDirectory(request.IntegrationName, true);
+            // move temp file to agent input directory
+            File.Move(tempFilePath, destinationFilePath);
 
-                // define new path to move temp file to
-                var destinationFilePath = Path.Combine(agentInputDirectory, $"{Guid.NewGuid()}.json");
-
-                // move temp file to agent input directory
-                File.Move(tempFilePath, destinationFilePath);
-
-                // return result
-                return new Response();
-            }
+            // return result
+            return new Response();
         }
     }
 }
