+++ Integration.Common/Requests/WaitForAgentOutputFile.cs
@@ -3,53 +3,52 @@ using Integration.Common.Extensions;
 using MediatR;
 using Microsoft.Extensions.Options;
 
-namespace Integration.Common.Requests
-{
-    public class WaitForAgentOutputFile : IRequest
-    {
-        public class Request : IRequest<string>
-        {
-            public string FileName { get; }
-            public string IntegrationName { get; }
+namespace Integration.Common.Requests;
 
-            public Request(string fileName, string integrationName)
-            {
-                FileName = fileName;
-                IntegrationName = integrationName;
-            }
+public class WaitForAgentOutputFile : IRequest
+{
+    public class Request : IRequest<string>
+    {
+        public string FileName { get; }
+        public string IntegrationName { get; }
+
+        public Request(string fileName, string integrationName)
+        {
+            FileName = fileName;
+            IntegrationName = integrationName;
+        }
+    }
+
+    public class Handler : IRequestHandler<Request, string>
+    {
+        private readonly IntegrationSettings _settings;
+
+        public Handler(IOptions<IntegrationSettings> options)
+        {
+            _settings = options.Value;
         }
 
-        public class Handler : IRequestHandler<Request, string>
+        public async Task<string> Handle(Request request, CancellationToken cancellationToken)
         {
-            private readonly AgentSettings _settings;
+            // define agent output directory
+            var agentOutputDirectory = _settings.GetOrCreateAgentOutputDirectory(request.IntegrationName, true);
 
-            public Handler(IOptions<AgentSettings> options)
+            // loop while cancel token is not cancelled and no files are found
+            while (!cancellationToken.IsCancellationRequested)
             {
-                _settings = options.Value;
+                // search files in output directory
+                var files = Directory.GetFiles(agentOutputDirectory, request.FileName);
+
+                // return response with files if any
+                if (files.Any())
+                    return files.First();
+
+                // delay before retry
+                await Task.Delay(500, cancellationToken).ConfigureAwait(false);
             }
 
-            public async Task<string> Handle(Request request, CancellationToken cancellationToken)
-            {
-                // define agent output directory
-                var agentOutputDirectory = _settings.GetOrCreateAgentOutputDirectory(request.IntegrationName, true);
-
-                // loop while cancel token is not cancelled and no files are found
-                while (!cancellationToken.IsCancellationRequested)
-                {
-                    // search files in output directory
-                    var files = Directory.GetFiles(agentOutputDirectory, request.FileName);
-
-                    // return response with files if any
-                    if (files.Any())
-                        return files.First();
-
-                    // delay before retry
-                    await Task.Delay(500, cancellationToken).ConfigureAwait(false);
-                }
-
-                // return empty response
-                return string.Empty;
-            }
+            // return empty response
+            return string.Empty;
         }
     }
 }
