+++ Integration.Host/appsettings.json
@@ -6,7 +6,7 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-  "AgentSettings":{
+  "IntegrationSettings":{
     "RootDirectory": "Path to agent directory - same as root directory value"
   }
 }
